requirement1.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
asgiref
attrs
backoff
bcrypt
build
cachetools
certifi
charset-normalizer
chroma-hnswlib
chromadb
click
coloredlogs
dataclasses-json
Deprecated
distro
durationpy
fastapi
filelock
flatbuffers
frozenlist
fsspec
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
grpcio
grpcio-status
h11
httpcore
httplib2
httptools
httpx
httpx-sse
huggingface-hub
humanfriendly
idna
importlib_metadata
importlib_resources
Jinja2
joblib
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
kubernetes
langchain
langchain-community
langchain-core
langchain-text-splitters
langsmith
markdown-it-py
MarkupSafe
marshmallow
mdurl
mmh3
monotonic
mpmath
multidict
mypy-extensions
networkx
numpy==1.24.4
oauthlib
onnxruntime
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-fastapi
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-util-http
orjson
overrides
packaging
pillow
posthog
propcache
proto-plus
protobuf
pyasn1
pyasn1_modules
pydantic
pydantic-settings
pydantic_core
Pygments
pyparsing
pypdf
PyPika
pyproject_hooks
python-dateutil
python-dotenv
PyYAML
referencing
regex
requests
requests-oauthlib
requests-toolbelt
rich
rpds-py
rsa
safetensors
scikit-learn
scipy
sentence-transformers
setuptools
shellingham
six
sniffio
SQLAlchemy
starlette
sympy
tenacity
threadpoolctl
tokenizers
torch
tqdm
transformers
typer
typing-inspect
typing-inspection
typing_extensions
uritemplate
urllib3
uvicorn
uvloop
watchfiles
websocket-client
websockets
wrapt
yarl
zipp
zstandard
flask